+++ src/index.ts
@@ -1,6 +1,6 @@
 // Onion Lasers Command Handler //
 export {Command, NamedCommand, RestCommand, CHANNEL_TYPE} from "./command";
-export {addInterceptRule} from "./handler";
+export {addInterceptRule, setExecuteCommandListener} from "./handler";
 export {launch} from "./interface";
 export * from "./lib";
 export {getCommandList, getCommandInfo} from "./loader";