+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#
+- Made `reactInOrder` public
+
 # 1.1.2
 - Fixed an issue where a bot would end up deleting all reactions regardless of whether or not the message had a reaction event listener to begin with (i.e. for `paginate`).